+++ src/main/java/com/example/comprehensivedegisn/batch/kakao_map/jibun/KaKaoMapBatchConfigurationWithJibun.java
@@ -2,12 +2,12 @@ package com.example.comprehensivedegisn.batch.kakao_map.jibun;
 
 import com.example.comprehensivedegisn.adapter.domain.Gu;
 import com.example.comprehensivedegisn.batch.CacheRepository;
-import com.example.comprehensivedegisn.api_client.KaKaoApiClientWithJibun;
+import com.example.comprehensivedegisn.api_client.kakao.KaKaoApiClientWithJibun;
 import com.example.comprehensivedegisn.batch.kakao_map.KaKaoMapBaseConfiguration;
 import com.example.comprehensivedegisn.api_client.ApiClient;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.ApartmentGeoRecord;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.LocationRecord;
-import com.example.comprehensivedegisn.batch.kakao_map.KaKaoRestApiProperties;
+import com.example.comprehensivedegisn.api_client.kakao.KaKaoRestApiProperties;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.TransactionWithGu;
 import com.example.comprehensivedegisn.batch.query_dsl.QuerydslNoOffsetIdPagingItemReader;
 import com.example.comprehensivedegisn.batch.query_dsl.expression.Expression;
