+++ src/main/java/com/example/comprehensivedegisn/batch/kakao_map/road_name/KaKaoMapBatchConfigurationWithRoadName.java
@@ -1,9 +1,9 @@
 package com.example.comprehensivedegisn.batch.kakao_map.road_name;
 
 import com.example.comprehensivedegisn.batch.CacheRepository;
-import com.example.comprehensivedegisn.api_client.KaKaoApiClientWithRoadName;
+import com.example.comprehensivedegisn.api_client.kakao.KaKaoApiClientWithRoadName;
 import com.example.comprehensivedegisn.batch.kakao_map.KaKaoMapBaseConfiguration;
-import com.example.comprehensivedegisn.batch.kakao_map.KaKaoRestApiProperties;
+import com.example.comprehensivedegisn.api_client.kakao.KaKaoRestApiProperties;
 import com.example.comprehensivedegisn.api_client.ApiClient;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.ApartmentGeoRecord;
 import com.example.comprehensivedegisn.batch.kakao_map.dto.LocationRecord;
