+++ src/main/java/com/example/comprehensivedegisn/batch/open_api/OpenApiBatchConfiguration.java
@@ -1,6 +1,6 @@
 package com.example.comprehensivedegisn.batch.open_api;
 
-import com.example.comprehensivedegisn.api_client.OpenApiClient;
+import com.example.comprehensivedegisn.api_client.open_api.OpenApiClient;
 import com.example.comprehensivedegisn.batch.open_api.dto.ApartmentDetailResponse;
 import com.example.comprehensivedegisn.adapter.repository.dong.QuerydslDongRepository;
 import lombok.RequiredArgsConstructor;
